Rename back to _debug_mode
