+++ dali/python/nvidia/dali/fn.py
@@ -76,7 +76,7 @@ def _wrap_op_fn(op_class, wrapper_name, wrapper_doc):
         return op_class(**init_args)(*inputs, **call_args)
 
     def fn_wrapper(*inputs, **kwargs):
-        from nvidia.dali.debug_mode import _PipelineDebug
+        from nvidia.dali._debug_mode import _PipelineDebug
         current_pipeline = _PipelineDebug.current()
         if getattr(current_pipeline, '_debug_on', False):
             return current_pipeline._wrap_op_call(op_class, *inputs, **kwargs)
